Create src/models folder and divided current models in separate files
